[iec] change receiveByte() to return uint8_t move #ifdef JIFFYDOS to disable check

+++ lib/bus/iec/protocol/cpbstandardserial.h
@@ -34,7 +34,7 @@ namespace Protocol
     class CPBStandardSerial : public IECProtocol
     {
         private:
-        int8_t receiveBits();
+        uint8_t receiveBits();
         bool sendBits(uint8_t data);
 
 
@@ -53,7 +53,7 @@ namespace Protocol
          * @brief receive byte from bus
          * @return The byte received from bus.
          */
-        virtual int16_t receiveByte();
+        virtual uint8_t receiveByte();
         
         /**
          * @brief send byte to bus
